Fix warnings 7/n - casts

+++ source/benchmarks/memory_benchmark/implementations/l0/usm_copy_multiple_blits_l0.cpp
@@ -97,7 +97,7 @@ static TestResult run(const UsmCopyMultipleBlitsArguments &arguments, Statistics
             blitSizeAssigner.addMainCopyEngine();
         } else {
             cmdQueueDesc.ordinal = linkCopyOrdinal;
-            cmdQueueDesc.index = blitterIndex - queueProperties[mainCopyOrdinal].numQueues;
+            cmdQueueDesc.index = static_cast<uint32_t>(blitterIndex - queueProperties[mainCopyOrdinal].numQueues);
             blitSizeAssigner.addLinkCopyEngine();
         }
